Use more targeted approach to suspending transient override
It is difficult to determine on `minibuffer-setup-hook' and
`minibuffer-exit-hook' where in the active transient command's
life-cycle the current minibuffer appears, and we failed to do
so in some edge-cases.

Avoid this complication by not adding any functions to these hooks
whenever a transient prefix command is setup.  Instead add simplified
functions right before using the minibuffer for the benefit of an
infix command and arrange for them to not affect any other minibuffer
usage, including (potentially indirect) recursive usage.